requirements.txt
requests
python-dotenv
pdf2image==1.17.0
aryn-sdk
anthropic